Create JATALA
FRESH MIMOSA HOSTILIS IN STOCK

We are proud to supply the United States with the best Mimosa Hostilis root bark. Therefore, we strongly believe that such materials should be available to all. Our priority is to meet every customers’ satisfaction and expectations. We aim to have 100% of our customers come back and be happy to order from us again. We are looking to distribute our product mainly throughout the U.S.A and Canada. Therefore, other foreign countries may not be eligible for export. However, please feel free to contact us for inquiries on shipment eligibility. We offer USPS free flat rate shipping to our U.S clientele. Shipping takes place within one to three days of the order’s placement. We have mimosa hostilis inner root bark. This is not only due to high demand but is also because we only serve fresh Mimosa Hostilis root bark to our clients. We are constantly receiving from Mexico and shipping that newly received bark to our customers in order to achieve the high quality that we are known for. Please wait a total of 12 hours before placing a request for shipping information. Our Mimosa Hostilis Bark comes from the Chiapas region of Mexico where it is harvested naturally by a local team.  The Mimosa Hostilis inner root bark has been personally harvested, treated, and packed. The soil contains no additional nutrients since Mimosa occurs naturally in this region. Comparatively, this makes our product the best available on the market. We do not include the trunk in our product and accordingly only believe in using the inner root bark. 
abraxasroots@gmail.com
https://abraxasroots.com/